added delete protection to original roles

+++ backend/app/GraphQL/Mutations/DeleteRole.php
@@ -2,6 +2,7 @@
 
 namespace App\GraphQL\Mutations;
 
+use App\Exceptions\CustomException;
 use App\Models\Role;
 use App\Models\User;
 
@@ -13,6 +14,9 @@ final class DeleteRole
      */
     public function __invoke($_, array $args)
     {
+        if ($args['id'] <= 3 && $args['id'] >= 1) {
+            return new CustomException('Cannot delete original roles');
+        }
         $role = Role::find($args['id']);
         User::where('role_id', $args['id'])->update(['role_id' => 3]);
         $role->delete();
